requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
alembic==1.13.1
